+++ ckanext/fork/util.py
@@ -1,4 +1,8 @@
+import logging
+import ckan.logic as logic
 from ckan.plugins import toolkit
+
+log = logging.getLogger(__name__)
 
 
 def get_forked_data(context, resource_id, activity_id=None):
@@ -49,3 +53,29 @@ def blob_storage_fork_resource(context, resource):
 
     resource['fork_activity'] = forked_data['activity_id']
     return resource
+
+
+def get_current_resource(context, resource):
+    current_resource = {}
+    if resource.get("id"):
+        try:
+            current_resource = toolkit.get_action('resource_show')(
+                context,
+                {"id": resource["id"]}
+            )
+        except logic.NotFound:
+            log.info(f"Resource {resource['id']} does not exist "
+                     "- must be creating a new resource with specific id.")
+    return current_resource
+
+
+def check_metadata_for_file_change(current, resource):
+    file_metadata_changed = False
+    if resource.get("fork_resource") or current.get("fork_resource"):
+        for key in ['lfs_prefix', 'size', 'sha256', 'url_type']:
+            new_value = resource.get(key, "")
+            if new_value:
+                original_value = current.get(key, "")
+                if original_value != new_value:
+                    file_metadata_changed = True
+    return file_metadata_changed
